Move antiraid to new trigger command system.
